+++ pkg/thumbnails/storage/filesystem.go
@@ -1,3 +1,3 @@
-package cache
+package storage
 
 // TODO: implement filesystem cache for longer persistence
