+++ pkg/thumbnails/storage/storage.go
@@ -1,11 +1,11 @@
-package cache
+package storage
 
 import (
 	"image"
 )
 
-// Cache defines the interface for a thumbnail cache.
-type Cache interface {
+// Storage defines the interface for a thumbnail store.
+type Storage interface {
 	Get(key string) image.Image
 	Set(key string, thumbnail image.Image) (image.Image, error)
 }
